Changed imgui image to cast to uintptr_t

+++ Engine/src/ImGui/ImGuiUtilities.cpp
@@ -18,7 +18,7 @@ void ImGui::Image(Ref<Texture> texture, const ImVec2& size, const ImVec4& tint_c
 {
 	if (texture)
 	{
-		ImTextureID my_tex_id = (void*)(uint64_t)texture->GetRendererID();
+		ImTextureID my_tex_id = (void*)(uintptr_t)texture->GetRendererID();
 		ImGui::Image(my_tex_id, size, ImVec2(0, 1), ImVec2(1, 0), tint_col, border_col);
 	}
 }
@@ -27,7 +27,7 @@ void ImGui::Image(Ref<Texture> texture, const ImVec2& size, const ImVec4& tint_c
 
 IMGUI_API bool ImGui::ImageButton(Ref<Texture> texture, const ImVec2& size, int frame_padding, const ImVec4& bg_col, const ImVec4& tint_col)
 {
-	ImTextureID my_tex_id = (void*)(uint64_t)texture->GetRendererID();
+	ImTextureID my_tex_id = (void*)(uintptr_t)texture->GetRendererID();
 	return ImGui::ImageButton(my_tex_id, size, ImVec2(0, 1), ImVec2(1, 0), frame_padding, bg_col, tint_col);
 }
